refactor: move apps files to cli
