+++ packages/cli/src/commands/apps/domains/clear.ts
