+++ packages/cli/src/commands/apps/domains/index.ts
