+++ packages/cli/src/commands/apps/domains/info.ts
