+++ packages/cli/src/commands/apps/domains/remove.ts
