+++ packages/cli/test/unit/commands/apps/domains/clear.unit.test.ts
