+++ packages/cli/test/unit/commands/apps/domains/info.unit.test.ts
