+++ packages/cli/test/unit/commands/apps/domains/remove.unit.test.ts
